put translations into separate module "lang.ts"
