+++ src/components/PlaceGroupsSelect.tsx
@@ -10,7 +10,7 @@ import ListItemText from "@material-ui/core/ListItemText";
 
 import { Dataset } from '../model/dataset';
 import { PlaceGroup } from '../model/place';
-import { I18N } from '../config';
+import { I18N } from '../lang';
 
 
 const styles = (theme: Theme) => createStyles(
